+++ src/2_setup/player_setup.c
@@ -6,7 +6,7 @@
 /*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/29 08:28:05 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/20 07:16:22 by fschuber         ###   ########.fr       */
+/*   Updated: 2024/06/26 14:05:00 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,4 +18,9 @@ void	setup_player(void)
 	player()->transform.pos.y = player()->spawn_transform.pos.y + 0.5;
 	player()->transform.rot.x = player()->spawn_transform.rot.x;
 	player()->transform.rot.y = player()->spawn_transform.rot.y;
+	player()->inv.keys = 0;
+	player()->inv.water_orbs = 0;
+	player()->inv.fire_orbs = 0;
+	player()->inv.earth_orbs = 0;
+	player()->inv.air_orbs = 0;
 }